+++ src/cli/pile/merge.rs
@@ -3,8 +3,11 @@ use std::collections::{HashMap, HashSet};
 use std::convert::TryInto;
 use std::path::PathBuf;
 
+use triblespace::prelude::blobschemas::LongString;
+use triblespace::prelude::BlobStore;
 use triblespace::prelude::BlobStoreGet;
 use triblespace::prelude::BranchStore;
+use triblespace::prelude::View;
 use triblespace_core::id::Id;
 use triblespace_core::repo::pile::Pile;
 use triblespace_core::repo::Repository;
@@ -13,8 +16,6 @@ use triblespace_core::value::schemas::hash::Blake3;
 use triblespace_core::value::schemas::hash::Handle;
 use triblespace_core::value::schemas::hash::Hash;
 use triblespace_core::value::Value;
-use triblespace::prelude::blobschemas::LongString;
-use triblespace::prelude::View;
 
 use super::signing::load_signing_key;